Adds upload attachment and uses it as profile picture upload. Also adds profile picture update
